requirements.txt
Django==3.2.5
djangorestframework==3.12.4
django-rest-auth==0.9.5